Refactor departures route
This started as an attempt to use the standard `model` hook with
`this.refresh()`. Unfortunately, that didn't work with needing to 
maintain the current list of departures while simultaneously fetching a 
new list. I just ended up removing the isLoading logic because we're 
relying on the isLoading.hbs template for that.