Add departureFetcher object for smarter polling
On mobile devices, JS execution is suspended on sleep, so when the 
device wakes up it could potentially have stale data on the screen until 
the next poll event occurs (as much as 15 seconds after wake). This 
commit does a setInterval at 0.5 seconds that checks if it needs to 
refresh.